Enum apiai::lang::Language

pub enum Language {
    BrazilianPortuguese,
    ChineseCantonese,
    ChineseSimplified,
    ChineseTraditional,
    English,
    Dutch,
    French,
    German,
    Italian,
    Japanese,
    Korean,
    Portuguese,
    Russian,
    Spanish,
    Ukranian,
}

Enum defines API.ai supported languages and their text values used on the server

Variants
BrazilianPortugueseChineseCantoneseChineseSimplifiedChineseTraditionalEnglishDutchFrenchGermanItalianJapaneseKoreanPortugueseRussianSpanishUkranian
Methods

impl Language

fn value(&self) -> &'static str

This function can be called by a language enum instance to get the associated string for use inside the api