Struct apiai::structure::ApiResponse

pub struct ApiResponse {
    pub id: String,
    pub timestamp: String,
    pub lang: Language,
    pub result: ApiResult,
    pub status: ApiStatus,
    pub session_id: String,
}

Response from API.AI call

Fields
id: String timestamp: String lang: Language result: ApiResult status: ApiStatus

The status object is returned with every request and indicates if the request was successful. If it is not successful, error information is included.

session_id: String

Session ID